lots of changes as we start to refactor
A few notes:
Webkit (not sure about FF) doesn't support Array keys, so we still need
two separate object stores. The plan now is for the 'ids' object store
to contain all the metadata about the revisions on disk and the 'revs'
object store to only contain document bodies for the revs in conflict
being kept around.

 * Start to move to Node.JS-style callbacks
 * makeCouch -> pouch.open
 * Started to refactor bulkDocs() as the central update logic
 * pouch.put() now uses pouch.bulkDocs() to do the work, as it should

Much more to do.