Totally busted merging of rev trees
If I don't push something @dietrich won't believe that I'm actually
doing this. So, here's the busted code.

Still to do:
Something's not quite right with the recursion in my head. There are
two phases here that aren't represented right.

1) Fold together trees as far as they are identical
2) Graft when they diverge

If phase (1) finds nothing identical, the result is two trees, i.e.
no merge occurred. If phase (1) finds _any_ bit that's the same then
phase (2) should always return one tree, possibly with branching in
it (unless it was a linear history that is simply updated).

Ideally, this routine should be able to return the new set of disjoint
trees, and all the leaves (in {start : X, ids : [...]) format (as in
?revs=true).

Anyway. Some progress is now pushed in case I get hit by a bus.