(#664) - fix traverseRevTree in visualizeRevTree
With rev_tree slight redesign traverseRevTree got changed but in
visualizeRevTree we got its copy as traverseRevTree was not exposed
as public API. Now it is but we probably don't plan to support it.

It is quite sad we don't have any tests for this plugin and it just
randomly broke.