(#909) - Enabled paging for replication of doc_id's
Currently the finished value is incorrectly set to true when there are
no results in the first 25 records
(CHANGES_LIMIT = 25) when filtering by a list of doc_ids. This stops
the loop and prevents the doc_id's from being replicated.
Created the new pagingCount field and set it to the remoteLastSeq
value. PagingCount is decremented by
CHANGES_LIMIT every time the fetched callback runs. The logic for
finished has been modified so that even if
resultsLength ===0, finished will be false if pagingCount > 0. This may
also work for other types of filtered replication.