(#1015) - Fixed some typos and commas
and made minor changes to wording

Just took a pass through the getting started page and corrected some missing apostrophes and other typos. I also changed the wording of a few sentences. More could be done here, but I tried to fix bits that looked quick and easy to fix.