(#1239) - Favor GETs over POSTs for Safari
The test "all_docs: http-1: Testing allDocs opts.keys" breaks
in Safari because Safari cannot comprehend why a POST
would return a 304.  This commit fixes that.

Currently we always turn a GET into
a POST if "keys" is provided. So my proposal is to only do this
if "keys" would make the URL longer than 2000 characters,
since CouchDB accepts both a GET and a POST.