Issue #1267 - complete before cancel on continuous changes
In some cases (perhaps limited to LevelDB), a continuous changes feed will
call the complete callback immediately after calling the onChange callback
for each change in the database at the time changes was started, then call
onChange for any subsequent changes, resulting in onChange being called after
complete. This is, at best, confusing and seems unnecessary: the continuous
change feed is not complete at the time the complete callback is called.

This change is to utils.processChanges, to make the call to opts.complete
conditional on opts.continuous not being true.