(#1981) - more reliable, greppable perf tests
Added the ability to specify ?grep=foo to limit
the perf tests to a single test/suite. Logged
errors to the document, so we know to disregard
the results.

Increased the iterations to a point where the tests
still complete in a reasonable amount of time even
in a slow browser, but aren't so fast that we can't
tell different browsers apart.  Still based on gut
feelings, not statistical measures of significance.