(#1990) - move allDocs keys query to adapter.js
This moves the allDocs() logic when opts.keys is
used to adapter.js, which allows to just do a
simple multi-key fetch for each non-http adapter.

This improves the performance of leveldb.js and should
not have an impact on the performance of websql.js or
idb.js.  Plus it's less code.