(#2035) - make the CORS proxy optional
The CORS proxy is really flakey, and it contributes to
our intermittent errors.  However, for reasons that
are mysterious to me, Saucelabs gives us an error
without it, on iPhone, Safari, and IE.

This adds an optional flag, CORS_PROXY=true, which
can be used to enable the proxy.  Otherwise we
set the CORS config on CouchDB and use 5984 directly.