Don't bail just because db._docCountQueue.docCount is -1
There's a condition when the queue length is non-zero, but the docCount is still -1 -- in that case applyNextDocCountDelta() just returns silently and the adapter never finishes initializing itself because its callback never fires.  This is just a quicky hack-fix, I'm not familiar enough with all the parts to know what other ramifications this might have.