(2143) Get PouchDB's scripts to run properly on Windows
bin/build-plugin.sh - The resolve library in browerify acts differently on
Linux and Windows when
resolving ./lib/deps/migrate. But getting rid of the path and just using
the package name migrate (since this is a rename) seems to work
in both places. Also got rid of the accidental duplication of the migrate
line.

package.json - For calls to sh scripts change the call from being a path
 starting with './bin/' to start with 'sh bin/' and change calls to js
 files to change from './bin/' to 'node ./bin/'. Also had to set
 run-test.sh to use 'bash' and not 'sh' because it's a bash script.

test.attachments.js - Added the hashes for the attachments generated
by CouchDB on Windows. For some reason they are different than those
on Linux. We really should remove these hashes all together, please
see #3156