(#136) - Fix "Test create attachment without callback" against clustered CouchDB
We cannot rely on an ordered update sequence in clustered CouchDB.
Instead, request the _changes feed following the initial docment update
using the "since" parameter. There is still potential for this to fail
because _changes may return entries before that specified by "since" -
the guarantee is that all entries following "since" will be returned.

Possibly one to review if we find that the test fails occasionally
(for me it is stable).