(#136) - Test against 1 node CouchDB cluster
When spinning up a CouchDB cluster to test against, just use one node.
This ensures that we're still testing against the clustered interface but
removes any problems with eventual consistency between nodes in the cluster.

I'd suggest reverting this once we have all the tests passing but, for now,
this is useful to remove much of the "noise" in the failures.