(#136) - Ignore "changes last seq" against CouchDB 2.0
In clusterd CouchDB, the sequence number returned to the client is an
encoded representation of the various shard states. Whilst the last
"seq" value in _changes.results and the "last_seq" field should evaluate
to the same thing on the server, their encoded representation may differ.
Therefore, the assertions in this test don't hold for clustered CouchDB.