(#3286) - test in Chrome 36
Chrome 36 had no blob support, 37 was when they added broken
support, and 38 was when it was finally fixed.

I would like to test in all three. Chrome is the most popular
desktop browser on the planet, and it's also very close
to what mobile Android users experience. Let's play it safe
and test all three versions.