(#136) - Fix "Reporting write failures (#942)"
Change the "Reporting write failures (#942)" test to avoid dependencies
on the _changes feed order. This was failing against CouchDB 2.0 because
sometimes document "b" would come before "a" in the feed. The mocking
function would return a write error for document "b" which caused the
replication to abort at which point the assertions around documents
read / written fail.

The mocked bulk_docs endpoint now returns failure for the second
document, regardless of the _id.