(#2632) - remove browserified Buffer from build
So apparently Browserify does this thing where,
if you do "./mydep": "false", you need to reference
all possible relative paths used to that dependency,
whereas if you just swap it out for an empty
implementation then it works regardless of the relative
path used.

This solution is dumb, but it cuts down the min+gz size
from 51185 to 46298.