(#136) - Do not use sequence number comparison during replication
In CouchDB 2.0 / Cloudant, sequence numbers are non-integer values -
comparison (and even equality checks) between them doesn't make sense.

Prior to this commit, the replicator would fetch more changes based on
a comparison between the since parameter and the last_seq value returned
from the changes feed. In the new world, the only sure way of knowing
whether replication is complete is to check that no results were
returned from a call to _changes.

Unfortunately, the replicator doesn't have visibility into the raw
_changes feed result and client-side filtering could mask the fact that
results were returned. This commit prevents a filter being used directly
with the changes feed via the replicator and instead moves the filtering
logic to the replicator.

Whilst the tests pass, this changes the semantics of the onChange event
when using the replicator. Previously, onChange fired for documents
which passed the client-side filter (if specified) - after this commit
it fires for all documents (because the filtering logic is now in the
replicator).