(#3395) - Return 404 when open_revs is requested for missing doc
CouchDB 1.X crashes and returns a 200 response when a request for
open_revs on a non-existent document is made. In CouchDB 2.0 this is
fixed (COUCHDB-2517) and it now returns a 404 which seems like the
correct response.

We were previously matching the CouchDB 1.X behavior and hiding this
error condition. This commit allows the error to propagate up, matching
the new CouchDB behaviour.