(#3220) - Test which illustrates sync/async on('change')
In the current implementation, db.on('change') behaves differently
depending on the database adapter. In Chrome/WebSQL the call is
synchronous - when it returns the changes feed will be listening
(from current seq - 1) and so it fails because it triggers a change
event for the document which already exists in the db.

In FF/IndexedDB, the call is async. In this test, we PUT a document
after setting up the changes listener. The test passes because the PUT
completes before the changes listener starts. As mentioned above,
the changes listener emits an event for the latest document updated
in the database (due to the update_seq - 1 logic) and so this passes.