(#3220) - Fix test.events on('change') tests
db.on('change') is async. In the event emitter tests, we PUT a document
after setting up the changes listener. The existing test previously
passed because the PUT completed before the changes listener started.
The changes listener emitted an event for the latest document updated
in the database (due to the update_seq - 1 logic) and so assertion was
correct. In the event that the changes feed began listening before
the PUT, the update_seq was normalized to zero so again, it passed.

This commit adds a second test which demonstrates where this breaks down
- it fails when the -1 logic is used to compensate for the async
behaviour.

The correct fix is to recognise the async nature of the on('changes')
and have the tests wait until a change they recognise is emitted.