(#2674) - WIP: support atts_since parameter
Work in progress to support the atts_since parameter.

For the HTTP adapter, just pass through to the external target. This
commit also adds support for atts_since to the the idb and leveldb
adapters - we need to store an additional piece of metadata (revpos)
alongside attachments so there is a schema change. However, existing
attachments without the revpos should still work - the response to
atts_since will just be suboptimal.

TBD whether it's worth using a migration to set a default revpos locally
(I think it could default to the current revpos for the document).

TODO:
- add support for websql, sqlite
- use atts_since during replication