(#3552) - Make a test pass for Sync Gateway
This makes test.replication.js-local-http by skipping assertions
that expect database info to contain a doc count. If there is a
doc count we make sure it's correct.

Also Sync Gateway sends an additional _revisions field which was
messing up deep equal, so instead now I'm testing directly for
the field we expect.