(#3561) - verifyInfo for CSG
This duck types db infos from Sync Gateway and makes allowances
for things like access control sequences and lack of doc_count.

We don't switch on isSyncGateway() explicitly b/c we don't want
to drop these assertions for the whole environment, just when
the info is from Sync Gateway. If we want to make sure it's not
being lenient in other environments we could refactor the verifyInfo
function to look more like a switch statement. Maybe the time to
do this is on the NEXT environment that gets added. :)