(#3689) - Add ajax to GET and fix npm run dev on Windows
Needed the ability to actually use the advertised support for ajax on GET
requests.

test.headers.js - Fixed it up to use promises for the tests. This was
important for after() because otherwise for some reason done never
got called. Also added tests for new GET functionality.

npm run dev - The package.json was edited to use an environment variable
to set the client variable to 'dev'. But this doesn't work on Windows.
So I edited bin/run-test.sh to support an optional command line argument
and then edited package.json to use that argument to pass in 'dev'