(#3688) - correctly es3ify
es3ify is not really doing what we want it to do. When used as a
browserify transform, it is unsafe, because it only applies to
OUR OWN code, not necessarily dependencies.

Furthermore, it's been causing problems from day one due to weirdnesses
where projects that depend on PouchDB and browserify it themselves
also need to require es3ify directly. That's not cool.

My fix is to use es3ify in the same way that we use derequire - i.e.
just as a command-line tool when building, which runs over the entire
file.

I've manually confirmed that this outputs es3ified code and fixes #3688.