(#3723) - add a "who's using" page
I think this is one of the first questions people
ask themselves when they're evaluating a piece of
software, so it's important for us to market ourselves.

The images were scaled down a bit to fit well on the page.
Feel free to tinker if you think it needs it @nickcolley.

Also we should definitely tweet and ask people to submit their
projects. I already emailed Chilean fishing boat folks to see
if we could get a statement from them. :)