(#3424) - implement bulkGet()
Related to pouchdb/express-pouchdb#196 and
apache/couchdb-chttpd#33

This adds a `bulkGet()` function to all PouchDB
objects, which in the case of local databases runs
a shim that simply collates a bunch of `get()` requests.
This logic is in `bulk-get-shim.js`.

In the case of the http adapter, the client checks if
the server responds to `_bulk_get` with a 40x error. If
no error is returned, then the server is assumed to
implement `_bulk_get` and that API is used. Else the shim
is used.

I have tested this against CouchDB 1.6 and PouchDB Server
(82b297e) with 100% success in both. Also manual testing
showed that the `_bulk_get` API was used for PouchDB Server
but not CouchDB.

Three cheers for bulk replication!