(#3699) - fix retry memory leaks
Goals:
* Fixes #3699 by ensuring that listeners aren't leaked (heavily tested)
* The tests should fail instantly rather than time out

You can verify that the tests fail instantly when we have a memory
leak, by running the same tests against master:

    git checkout origin/master -- lib/

You'll see that we've been leaking the following events:

  * destroyed, cancel, complete, error

There are some events we weren't leaking, but I added tests for them
anyway:

  * paused, active, change

The reason the tests are slow is because we have no way to configure
the backoff (currently a random number between 0 and 2000 ms).
Once we add a way to configure the backoff, we can make these tests
faster.