(#3824) - fix live replication batch size
Fixes #3824 by actually using live changes from the get-go. Before,
we would do `changes()` once without `live`, then once again
without `live`, then finally the third time with `live`, leading
to the bug.

The reason this is untested is because it would be pretty much
impossible to write a test for. The only test I can think of would
be to check the exact number of times that `changes()` was called during
a live replication, but that seems flakey.

However, if you check the network tab during live replication, you'll
see we're doing far fewer HTTP requests than before. So this is a
win for replication performance (again).