(#3858) - fix all EventEmitter leaks in tests
This also adds a new module I wrote which throws an error
if the maximum number of listeners is ever exceeded. It
only runs in the Node tests, but this should be a great
way to prevent future memory leaks.

I can confirm that the module works, and that these fixes were
actually necessary to get the tests passing again.