(#4070) - Fix check for localstorage in Safari private browsing mode
Safari private mode still exposes the localStorage object but sets the DOM quota to 0, making it non-functional.

I'd like to still be able to use the pouchdb memory adapter in safari private mode, but this test seems to make it fail, since it attempts to use localStorage thinking it will be functional.