(#3921) - conditional JSON.parse
Only runs `vuvuzela.parse` if the string is very large.
This function seemed to be a big offender in the profiler,
probably due to the deoptimization of try/catch, so it can
only be helpful to remove it where possible.

If this passes, I'll try to get some perf numbers.