(#4530) - allow bulkGet requests without an explicit rev
If no revision is specified for a document when using the bulkGet API,
fetch the implicit winning revision. Rather than making an additional
request to fetch the rev, we depend on the implicit behaviour of db.get
without an open_revs parameter.

This commit also fixes a couple of documentation issues discovered
whilst writing the tests.