(#4642) - deterministic revision numbers
Naive implementation of a deterministic revision number. As a first
pass, this uses the md5 hash of the document to generate a new revision,
only in the case that an existing _rev is present (i.e. for doc updates).

The current structure of the document update logic makes it tricky to
incorporate the full rev history, hence we fall back to uuid-based revs
for new documents or revisions following a deletion. A possible
workaround might be to explicitly attach any existing _rev before the
document update logic is called.

This patch still makes client side conflict resolution easier - if
conflicts are resolved concurrently then they should generate the same
rev tree.