(#4712) - return conflict error when inserting using unknown _rev
When inserting a new document with a specific _rev, we previously
allowed the insertion instead of generating a conflict error
(as CouchDB does).

It was a bit fiddly to figure out where to add this check but I settled
on processDocs.js. This commit adds a test that when inserting a document
new to the database (and new_edits is true), the root of the revision tree
must not be missing. Normally, PouchDB will generate a new, available
root node so the lack of this indicates that the user specified a _rev
that the database doesn't recognise.

Updates to a document which the database already knows about will not
flow down this code path - instead they go to updateDoc.js which performs
its own conflict test.