"views" in design documents take a map function
Before, the above code caused couchDB to return an error http status code.

This is because the example map function was not in the "map" key in the design
document. Therefore Couch saved a view without a map function, aka the
`views` property when querying couch for the document showed just an empty
object.