Fix for triple request with replication
processPendingBatch is called with immediate at true in the onChange
callback when batches is empty, which is unnecessary and makes no real
sense.
It is necessary for now as a side effect to trigger batch
startNextBatch()

I added a simple flush argument to onChange for the http adapter that
indicates if a change is the last one available and should therefor
trigger an immediate new batch rollup.