(#4764) - Improve replication for read-only source databases
This improves detection of read-only source databases. Before this
change, there were two different checks for read-only source databases.
One version was checking for `4xx` response status codes when trying to
update the replication log in the source database. The other version only
checked for a `401` status code when trying to create a new replication
log in the source database. We now always check for any client error
status code (`4xx`). This avoids problems when the source database
responds with client error status codes other than `401` to replication
log write requests.

This commit also adds an early return if the source database is
read-only. The goal is to avoid that the checkpointer unnecessarily
reattempts reading and writing checkpoints.