(#4748) - skip bulk_get detection for Cloudant
A bit of a hacky fix. Cloudant always sets instance_start_time to "0"
so use this to shim bulk_get support. A request to /_bulk_get on a
Cloudant database today results in a 401 response with a
WWW-Authenticate: Basic header, generating a browser authentication
dialogue.

Remove this when Cloudant supports _bulk_get (or can avoid the
WWW-Authenticate response header for it).