(#2709) - support `revpos` in attachment stubs (#5037)
Add the `revpos` property to attachment stubs on save. `revpos` is the doc's
`rev` prefix (a sequential number) from when the attachment was added or
updated. It is used during replication in the CouchDB replicator (and others
including CouchBase Lite) to avoid fetching attachments that have not changed
since last replication.

This solves a problem with attachments created in PouchDB, which are then
replicated to an Apache CouchDB instance. They end up having a `revpos` property
set to `0`. When the doc is later pull replicated from an iOS application using
CouchBase Lite it is skipped and does not get replicated. This is because CBL
rejects attachments with invalid attachment stub (see
https://github.com/couchbase/couchbase-lite-ios/issues/1200). `revpos:0` is
considered invalid in CBL.

Also this improves replication performance of documents with attachments created
by PouchDB, when they are later replicated using Apache CouchDBs own replicator.
The Apache CouchDB replicator itself handles `revpos:0` docs without a problem.
But the replication of those docs cannot be optimized; the attachments get
copied over again for every revision, even if they did not change.