(#5238) - Fix unhandled rejection in changes
changes - Checks to make sure someone is listening for the error event
before emitting it. See #5238 for details.

index - Registering listenerCount

listenerCount - Supports both the modern and legacy versions of
  listenerCount

test.changes.js - This test will fail if there is an unhandled rejection
so we use it to test that the fix worked.

utils - Enables adding/removing events on the either the process or
windows object depending on environment. And dealing with the fact
that the unhandledRejection event is spelled differently in node vs
the browser. Note that this is a NOOP in Firefox but works in Chrome. This
just means if there is a bug in Firefox it won't be caught.

* Fix @daleharvey 's comments

test.changes - Switched from using after to using a final catch all then
clause to remove the listener. Dale said this is closer to the style that
PouchDB uses.

utils - Provided wrappers specifically for unhandled rejection events per
Dale's comment