(#5272) fix checkpointing against Cloudant/CouchDB 2.0
An attempt to provide a better fix for #5145. It seems like
it's safe to eagerly write a checkpoint in the event there are no
changes returned (e.g. because a filter did not generate results).

This also removes some assertions around last_seq / since values
in the tests, since they don't hold in CouchDB 2.x / Cloudant.

This at least allows the tests to pass - possibly this means we
need more tests to cover situations where a small number of docs
are returned.