(#5379) do not update checkpoint if last_seq has not changed
Now that we checkpoint on every batch of changes (rather than when
replication is complete), avoid updating the checkpoint
when the source database has not changed. As a first pass, we determine
this based on whether _changes/last_seq progresses, though this won't
work if the source database is CouchDB 2.0 or Cloudant where sequence
values may change even if the underlying data doesn't. I can't think
of a way around this, but it's also an edge case - it just means that
a PouchDB database might grow if compaction isn't run.