(#5379) - do not update checkpoint if last_seq has not changed
Avoid updating the checkpoint when there is no change to the sequence
number. This prevents unnecessary updates to the checkpoint document,
generating traffic and compaction bloat.

This approach doesn't help if the source database is CouchDB 2.0 or
Cloudant because sequence values may change even if the underlying
database doesn't. I can't think of a way around this, but it's also an
edge case - it just means that a PouchDB database might grow if
compaction isn't run.