(#5172) - Differentiate between validation and non-validation errors
Up to now, all 401 and 403 error were treated the same, i.e. it was
assumed that it may be connected to a result of a validation function on
the destination database or to error saving a design as non-admin user
(there is also a case of saving checkpoint on a read-only source
database, but it is dealt with by the Checkpointer).

This change refactors error handling in replication, making failure that
results in aborting the replication more explicit. In the result, it is
now possible to accummulate errors that are not failures (i.e. do not
result in restarting the replication) in `result.errors`.

It also introduces a major in change in that it disables backoff for
authorization and authentication errors.

Finally, it sticks to the former logic that non-authorization error
encountered in `bulkDocs` response effects in a replication failure.
To my knowledge, the only error that we can encounter this way
(request succeeds, but response contains errors) is a document update
conflict. CouchDB seems to treat all the errors the same, i.e. it
reports them and goes forward [1][2].

[1] https://github.com/apache/couchdb/blob/1.6.x/src/couch_replicator/src/couch_replicator_api_wrap.erl#L381
[2] https://github.com/apache/couchdb/blob/1.6.x/src/couch_replicator/src/couch_replicator_worker.erl#L450