(#5612) - change view & filter function execution
This changes both `evalView` and `evalFilter` in pouchdb-core to enforce
strict mode. When running Node.js, we're furthermore executing the
respective function in a sandbox now. We're using the `vm` module
provided by Node.js for that.

BREAKING CHANGE: this may break view or filter functions, as we're now
generally enforcing strict mode for view and filter functions.